requirements.txt
paramiko==3.4.0
PyYAML==6.0.1
rich==13.7.0
python-dotenv==1.0.0
